requirements.txt
pillow~=9.0.0
numpy~=1.22.1
matplotlib~=3.5.1